requirements.txt
datasets==2.14.6
pandas==2.1.1
loguru==0.7.2
transformers==4.34.1
peft==0.5.0
accelerate==0.24.1
loralib==0.1.1
evaluate==0.4.1
rouge-score
python-dotenv==1.0.0
sentencepiece
protobuf
langchain==0.0.345
jinja2==3.1.2
strictyaml==1.7.3
openai==1.3.9
pint==0.23
elasticsearch==8.9.0
sentence_transformers==2.2.2